Update use-cases.md
Added in Wyre's proposed usage of the ERC-725, and ERC-735 implementations.

We see these as great tools for verification & compliance. While they're labeled as identity, we tend to separate the use-cases of identity into two definitions.

*Identity as a verification tool - E.g. Uploading government documentation etc...)
*Identity as a tool for SSI - E.g. I am proving that I am myself.

As I mentioned in the Telegram chat @tyleryasaka - We're pumped on 725/735 speaks well too the verification use-case :) Although, we felt that the cost of creating a personal 725 contract was a bit more expensive than needed in terms of gas/tx requirements, and also that there were some privacy considerations that might be ok for verification but not entirely necessary for SSI usage. Our ERC-DeFi is specifically targeted for _regulated_ entities so everyone can utilize this on-chain reservoir of verifications to essentially provide a compliant user-base for the person deploying in their dorm room, from the second they deploy their DeFi offering. I cover a lot of this in the repo highlighted above if you're interested to understand the differentiators etc... Believe me, we're not aiming to do more work than we need to. Haha :)

Happy to share more or answer questions etc as they may come up!

Thanks so much for all the hard work and I'm pumped to make sure we're supporting in every way possible through implementing, encouraging partner to do the same, giving critique to usage pro's/con's, etc... Seriously appreciate it @tyleryasaka  (and others involved who I don't know!)

👍 
🕵️ 
🕵️‍♀️ 
🆔
🎟 

🙌 🏗 🚀